Bug 1865304 - Step 5. Move the `intent` `'shared' functions to class is responsible for delegating delegating to the appropriate functions
